requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.1.0
audioread==2.1.6
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
editdistance==0.5.3
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
isort==4.3.4
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
mypy==0.670
mypy-extensions==0.4.1
networkx==2.2
numba==0.43.1
numpy==1.16.1
pandas==0.24.1
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pyaml==18.11.0
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
PyYAML==5.1
resampy==0.2.1
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
toolz==0.9.0
torch==1.1.0
torchvision==0.2.2.post3
tqdm==4.31.1
typed-ast==1.3.1
Werkzeug==0.15.2
wrapt==1.11.1
